removed compile warning
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@18218 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_lib/drt_condition_utils.cpp
@@ -190,7 +190,7 @@ void DRT::UTILS::FindConditionObjects(const DRT::Discretization& dis,
 {
   FindConditionedNodes(dis, conds, nodes);
 
-  for (unsigned i = 0; i < conds.size(); ++i)
+  for (size_t i = 0; i < conds.size(); ++i)
   {
     // get this condition's elements
     std::map< int, RCP< DRT::Element > >& geo = conds[i]->Geometry();
